% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoPart_MultiMS.R
\name{make_meta}
\alias{make_meta}
\title{Subdivide data into metadata columns only}
\usage{
make_meta(mm, use_cols)
}
\arguments{
\item{mm}{data frame of metadata and intensities as a signle data frame}

\item{use_cols}{column numbers to subset and return,
no range checking on the column
indeces is performed}
}
\value{
m_ints data frame of intensities only
}
\description{
Subdivide a data frame of protein metadata and intensities
into a data frame of meta data only
}
\examples{
data(mm_peptides)
head(mm_peptides)
metaCols = 1:7 # reusing this variable
m_prot.info = make_meta(mm_peptides, metaCols)
}
