% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenMS.R
\name{makeLMFormula}
\alias{makeLMFormula}
\title{String linear model formula suitable}
\usage{
makeLMFormula(eff, var_name = "")
}
\arguments{
\item{eff}{treatment group ordering for all samples being anlysed.
Single factor with 2+ teatment groups.
Used to generate formula and contrasts for lm().}

\item{var_name}{string variable name to use in the formula}
}
\value{
data structure with linea model formula and contrasts
 \describe{
  \item{lm.formula}{Lienar model formula suitable for right hand side of '
        ~' in lm(), ~ is not included int eh formula}
  \item{lm.params}{contrasts for lm(), here sum-to-zero constraint only}
}
}
\description{
Makes a string linear model formula suitable for the right hand side of the
equasion passed into lm()
}
\details{
eig_norm1 and eig_norm2
Here we incorporate the model matrix from EigenMS
normalization to find the significant
trends in the matrix of residuals.
}
\examples{
grps = as.factor(c('CG', 'CG', 'CG', 'mCG', 'mCG', 'mCG'))
makeLMFormula(grps, 'TREATS')
}
